MMA-2878: Sonar

+++ journalpost/src/main/java/no/nav/dokarkiv/journalpost/v1/services/SlettLogiskVedleggService.java
@@ -16,9 +16,12 @@ public class SlettLogiskVedleggService {
     }
 
     public void slettLogiskVedlegg(String dokumentInfoId, String logiskVedleggId) {
-        skannetInnholdRepository.findSkannetInnholdBySkannetInnholdIdAndDokumentinfoId(logiskVedleggId, dokumentInfoId)
-                .orElseThrow(() -> new LogiskVedleggIkkeFunnetException(String.format("Kunne ikke finne logisk vedlegg med logiskVedleggId=%s og dokumentId=%s i Joark", logiskVedleggId, dokumentInfoId)));
-
-        skannetInnholdRepository.deleteSkannetInnholdBySkannetInnholdIdAndDokumentinfoId(logiskVedleggId, dokumentInfoId);
+        if (!skannetInnholdRepository.findSkannetInnholdBySkannetInnholdIdAndDokumentinfoId(logiskVedleggId, dokumentInfoId).isPresent()) {
+            throw new LogiskVedleggIkkeFunnetException(
+                    String.format("Kunne ikke finne logisk vedlegg med logiskVedleggId=%s og dokumentId=%s i Joark", logiskVedleggId, dokumentInfoId)
+            );
+        } else {
+            skannetInnholdRepository.deleteSkannetInnholdBySkannetInnholdIdAndDokumentinfoId(logiskVedleggId, dokumentInfoId);
+        }
     }
 }
